+++ .github/workflows/android-build.yml
@@ -59,9 +59,6 @@ jobs:
       - name: Install app dependencies
         run: pnpm install
 
-      - name: Build Submodules
-        run: pnpm run build-submodules
-
       - name: Build SolidJS app
         env:
           VITE_NETWORK: signet
