+++ .github/workflows/playwright.yml
@@ -44,9 +44,6 @@ jobs:
       - name: Install app dependencies
         run: pnpm install
 
-      - name: Build Submodules
-        run: pnpm run build-submodules
-
       - run: npx playwright install --with-deps
         if: steps.playwright-cache.outputs.cache-hit != 'true'
       - run: npx playwright install-deps
