--- packages/ui/.gitignore
@@ -1,28 +1,0 @@
-
-dist
-.solid
-.output
-.vercel
-.netlify
-netlify
-
-# Environment
-.env
-.env*.local
-
-# dependencies
-/node_modules
-
-# IDEs and editors
-/.idea
-.project
-.classpath
-*.launch
-.settings/
-
-# Temp
-gitignore
-
-# System Files
-.DS_Store
-Thumbs.db
