--- packages/ui/.prettierrc
@@ -1,9 +1,0 @@
-{
-  "trailingComma": "none",
-  "tabWidth": 4,
-  "semi": true,
-  "singleQuote": false,
-  "arrowParens": "always",
-  "printWidth": 80,
-  "useTabs": false
-}
