--- packages/ui/package.json
@@ -1,72 +1,0 @@
-{
-  "name": "@mutinywallet/ui",
-  "version": "0.0.0",
-  "private": true,
-  "license": "MIT",
-  "scripts": {
-    "build": "rollup -c",
-    "check-types": "tsc --noemit",
-    "eslint": "eslint src",
-    "lint": "pnpm run eslint && pnpm run check-types",
-    "format": "npx prettier --write \"src/**/*.{ts,tsx,js,jsx,json,css,scss,md}\"",
-    "check-format": "npx prettier --check \"src/**/*.{ts,tsx,js,jsx,json,css,scss,md}\"",
-    "storybook": "storybook dev -p 6006",
-    "build-storybook": "storybook build"
-  },
-  "type": "module",
-  "files": [
-    "dist"
-  ],
-  "types": "./dist/types/index.d.ts",
-  "main": "./dist/esm/index.js",
-  "module": "./dist/esm/index.js",
-  "exports": {
-    ".": {
-      "solid": "./dist/source/index.jsx",
-      "default": "./dist/esm/index.js"
-    },
-    "./style.css": "./dist/esm/style.css",
-    "./tailwind.config.cjs": "./tailwind.config.cjs"
-  },
-  "devDependencies": {
-    "@storybook/addon-essentials": "^7.2.2",
-    "@storybook/addon-interactions": "^7.2.2",
-    "@storybook/addon-links": "^7.2.2",
-    "@storybook/addon-styling": "^1.3.6",
-    "@storybook/blocks": "^7.2.2",
-    "@storybook/testing-library": "^0.2.0",
-    "@typescript-eslint/eslint-plugin": "^6.2.0",
-    "@typescript-eslint/parser": "^6.2.0",
-    "autoprefixer": "^10.4.14",
-    "eslint": "^8.46.0",
-    "eslint-import-resolver-typescript": "2.7.1",
-    "eslint-plugin-import": "2.27.5",
-    "eslint-plugin-prettier": "4.2.1",
-    "eslint-plugin-solid": "0.11.0",
-    "postcss": "^8.4.26",
-    "prettier": "^2.8.8",
-    "react": "^18.2.0",
-    "react-dom": "^18.2.0",
-    "solid-start-node": "^0.3.0",
-    "storybook": "^7.2.2",
-    "storybook-solidjs": "1.0.0-beta.2",
-    "storybook-solidjs-vite": "^1.0.0-beta.2",
-    "tailwindcss": "^3.3.3",
-    "typescript": "^4.9.5"
-  },
-  "dependencies": {
-    "@kobalte/core": "^0.9.8",
-    "@rollup/plugin-image": "^3.0.2",
-    "@solidjs/meta": "^0.28.5",
-    "@solidjs/router": "^0.8.2",
-    "@types/node": "^18.17.1",
-    "rollup": "^3.28.0",
-    "rollup-plugin-css-only": "^4.3.0",
-    "rollup-preset-solid": "^2.0.1",
-    "solid-js": "^1.7.9",
-    "solid-start": "^0.2.29"
-  },
-  "peerDependencies": {
-    "solid-js": "^1.7.9"
-  }
-}
