--- packages/ui/postcss.config.cjs
@@ -1,12 +1,0 @@
-module.exports = {
-  addons: [
-    '@storybook/addon-links',
-    '@storybook/addon-essentials',
-    '@storybook/addon-interactions',
-    '@storybook/addon-styling',
-  ],
-  plugins: {
-    tailwindcss: {},
-    autoprefixer: {},
-  },
-}
