--- packages/ui/src/components/index.tsx
@@ -1,2 +1,0 @@
-export * from "./LoadingSpinner";
-export * from "./ExternalLink";
