--- packages/ui/src/index.tsx
@@ -1,3 +1,0 @@
-import "./index.css";
-
-export * from "./components";
