--- packages/ui/tsconfig.json
@@ -1,34 +1,0 @@
-{
-  "compilerOptions": {
-    "allowSyntheticDefaultImports": true,
-    "esModuleInterop": true,
-    "target": "ESNext",
-    "module": "ESNext",
-    "moduleResolution": "node",
-    "jsxImportSource": "solid-js",
-    "jsx": "preserve",
-    "strict": true,
-    "types": ["solid-start/env", "node"],
-    "baseUrl": "./",
-    "paths": {
-      "~/*": ["./src/*"]
-    },
-    "skipLibCheck": true
-  },
-  "include": [
-    "src/**/*",
-    "tailwind.config.cjs",
-    ".eslintrc.cjs"
-  ],
-  "exclude": [
-    "node_modules",
-    "./node_modules",
-    "./node_modules/*",
-    "./node_modules/**/*",
-    "./node_modules/@types/node/index.d.ts",
-    "node_modules/.pnpm/solid-js@1.7.7/node_modules/solid-js/types/*",
-    "node_modules/.pnpm/solid-start@0.3.1_@solidjs+meta@0.28.5_@solidjs+router@0.8.2_solid-js@1.7.9_solid-start-node@0.3.0_vite@4.4.6/node_modules/solid-start/islands/*",
-    "build",
-    "dist"
-  ]
-}
