+++ src/components/layout/LoadingSpinner.tsx
@@ -9,7 +9,7 @@ export const LoadingSpinner = (props: { big?: boolean; wide?: boolean }) => {
             }}
         >
             <svg
-                class="w-8 h-8 mr-2 text-gray-200 animate-spin fill-m-red place-self-center"
+                class="mr-2 h-8 w-8 animate-spin place-self-center fill-m-red text-gray-200"
                 viewBox="0 0 100 101"
                 fill="none"
                 xmlns="http://www.w3.org/2000/svg"
